[refactor] refactor bench engine cli
