+++ lmcache/cli/commands/bench/engine_bench/interactive/state.py
@@ -3,7 +3,7 @@
 
 ``InteractiveState`` holds the partially-configured benchmark parameters,
 can be initialized from CLI args or a saved JSON file, and can be
-converted to the ``argparse.Namespace`` that ``_bench_engine()`` expects.
+converted to the ``argparse.Namespace`` that ``run_engine_bench()`` expects.
 """
 
 # Standard
@@ -210,7 +210,7 @@ class InteractiveState:
         return state
 
     def to_namespace(self) -> argparse.Namespace:
-        """Convert to an ``argparse.Namespace`` compatible with ``_bench_engine``.
+        """Convert to ``argparse.Namespace`` compatible with ``run_engine_bench``.
 
         Fills defaults for any unset items, then builds the namespace
         with the attribute names that ``parse_args_to_config()`` and
